Refactor base pipeline shared by estimators
