+++ src/augury/ml_estimators/stacking_estimator.py
@@ -6,55 +6,28 @@ import numpy as np
 import pandas as pd
 from sklearn.base import BaseEstimator
 from sklearn.pipeline import Pipeline, make_pipeline
-from sklearn.preprocessing import StandardScaler, OneHotEncoder
-from sklearn.compose import ColumnTransformer
+from sklearn.preprocessing import StandardScaler
 from sklearn.ensemble import ExtraTreesRegressor
 from mlxtend.regressor import StackingRegressor
 import statsmodels.api as sm
 
 from augury.sklearn import (
-    CorrelationSelector,
-    ColumnDropper,
     TeammatchToMatchConverter,
     EloRegressor,
     DataFrameConverter,
     TimeSeriesRegressor,
 )
-from augury.settings import TEAM_NAMES, ROUND_TYPES, VENUES, CATEGORY_COLS, SEED
+from augury.settings import SEED
 from augury.types import R
-from .base_ml_estimator import BaseMLEstimator
+from .base_ml_estimator import BaseMLEstimator, BASE_ML_PIPELINE
 
 np.random.seed(SEED)
 
 
-ELO_MODEL_COLS = [
-    "prev_match_oppo_team",
-    "oppo_prev_match_oppo_team",
-    "prev_match_at_home",
-    "oppo_prev_match_at_home",
-    "date",
-]
 DEFAULT_MIN_YEAR = 1965
 
 ML_PIPELINE = make_pipeline(
-    DataFrameConverter(),
-    ColumnDropper(cols_to_drop=ELO_MODEL_COLS),
-    CorrelationSelector(cols_to_keep=CATEGORY_COLS),
-    ColumnTransformer(
-        [
-            (
-                "onehotencoder",
-                OneHotEncoder(
-                    categories=[TEAM_NAMES, TEAM_NAMES, ROUND_TYPES, VENUES],
-                    sparse=False,
-                    handle_unknown="ignore",
-                ),
-                CATEGORY_COLS,
-            )
-        ],
-        remainder=StandardScaler(),
-    ),
-    ExtraTreesRegressor(random_state=SEED),
+    DataFrameConverter(), BASE_ML_PIPELINE, ExtraTreesRegressor(random_state=SEED)
 )
 
 ELO_PIPELINE = make_pipeline(
@@ -119,7 +92,7 @@ class StackingEstimator(BaseMLEstimator):
                 )
 
         self.pipeline.set_params(
-            **{"pipeline-1__correlationselector__labels": y_filtered}
+            **{"pipeline-1__pipeline__correlationselector__labels": y_filtered}
         )
 
         return super().fit(X_filtered, y_filtered)
